requirements.txt
# requirements.txt

google-cloud-aiplatform[adk]
fastapi
uvicorn